chatroomList

+++ backend/src/main/java/com/team/HoneyBadger/Controller/ChatroomController.java
@@ -2,6 +2,7 @@ package com.team.HoneyBadger.Controller;
 
 import com.team.HoneyBadger.DTO.ChatroomRequestDTO;
 import com.team.HoneyBadger.DTO.ChatroomResponseDTO;
+import com.team.HoneyBadger.DTO.TokenDTO;
 import com.team.HoneyBadger.Entity.Chatroom;
 import com.team.HoneyBadger.Exception.DataDuplicateException;
 import com.team.HoneyBadger.Exception.DataNotFoundException;
@@ -21,34 +22,54 @@ import java.util.List;
 public class ChatroomController {
     private final MultiService multiService;
 
-    @GetMapping
-    public ResponseEntity<?> getChatroom(@RequestHeader Long chatroomId) { //채팅방 가져오기(찾아오기)
-        try {
-            ChatroomResponseDTO chatroomResponseDTO = multiService.getChatRoom(chatroomId);
+    @GetMapping("/list")
+    public ResponseEntity<?> getChatroomList(@RequestHeader("Authorization") String accessToken) { //채팅방 가져오기(찾아오기)
+        TokenDTO tokenDTO = multiService.checkToken(accessToken);
+        if (tokenDTO.isOK()) try {
+//            tokenDTO.username();
 
+            List<ChatroomResponseDTO> chatroomResponseDTOList = multiService.getChatRoomListByUser(tokenDTO.username());
+            return ResponseEntity.status(HttpStatus.OK).body(chatroomResponseDTOList);
+        } catch (DataNotFoundException ex) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(ex.getMessage());
+        }
+        else return tokenDTO.getResponseEntity();
+    }
+
+    @GetMapping
+    public ResponseEntity<?> getChatroom(@RequestHeader("Authorization") String accessToken, @RequestHeader Long chatroomId) { //채팅방 가져오기(찾아오기)
+        TokenDTO tokenDTO = multiService.checkToken(accessToken);
+        if (tokenDTO.isOK()) try {
+//            tokenDTO.username();
+
+            ChatroomResponseDTO chatroomResponseDTO = multiService.getChatRoom(chatroomId);
             return ResponseEntity.status(HttpStatus.OK).body(chatroomResponseDTO);
         } catch (DataNotFoundException ex) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).body(ex.getMessage());
         }
+        else return tokenDTO.getResponseEntity();
     }
 
     @PostMapping
-    public ResponseEntity<?> create(@RequestBody ChatroomRequestDTO chatroomRequestDTO) { //채팅방 만들기 (최초 생성 : 채팅방 이름, 참여자 이름 String)
-        try {
-            ChatroomResponseDTO chatroomResponseDTO= multiService.getChatRoomType(chatroomRequestDTO);
-            if (chatroomResponseDTO != null){
+    public ResponseEntity<?> create(@RequestHeader("Authorization") String accessToken, @RequestBody ChatroomRequestDTO chatroomRequestDTO) { //채팅방 만들기 (최초 생성 : 채팅방 이름, 참여자 이름 String)
+        TokenDTO tokenDTO = multiService.checkToken(accessToken);
+        if (tokenDTO.isOK()) try {
+            ChatroomResponseDTO chatroomResponseDTO = multiService.getChatRoomType(chatroomRequestDTO);
+            if (chatroomResponseDTO != null) {
                 return ResponseEntity.status(HttpStatus.OK).body(chatroomResponseDTO);
-            }else {
+            } else {
                 return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
             }
         } catch (DataNotFoundException ex) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).body(ex.getMessage());
         }
+        else return tokenDTO.getResponseEntity();
     }
 
     @PutMapping
-    public ResponseEntity<?> updateName(@RequestHeader Long chatroomId, @RequestBody ChatroomRequestDTO chatroomRequestDTO) { //채팅방 이름 수정
-        try {
+    public ResponseEntity<?> updateName(@RequestHeader("Authorization") String accessToken, @RequestHeader Long chatroomId, @RequestBody ChatroomRequestDTO chatroomRequestDTO) { //채팅방 이름 수정
+        TokenDTO tokenDTO = multiService.checkToken(accessToken);
+        if (tokenDTO.isOK()) try {
             ChatroomResponseDTO chatroomResponseDTO = multiService.updateChatroom(chatroomId, chatroomRequestDTO);
             return ResponseEntity.status(HttpStatus.OK).body(chatroomResponseDTO);
         } catch (DataNotFoundException ex) {
@@ -56,16 +77,19 @@ public class ChatroomController {
         } catch (IllegalArgumentException ex) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ex.getMessage());
         }
+        else return tokenDTO.getResponseEntity();
     }
 
     @DeleteMapping
-    public ResponseEntity<?> delete(@RequestHeader Long chatroomId) { //채팅방 삭제(참여자 테이블에서도 삭제됨.)
-        try {
+    public ResponseEntity<?> delete(@RequestHeader("Authorization") String accessToken, @RequestHeader Long chatroomId) { //채팅방 삭제(참여자 테이블에서도 삭제됨.)
+        TokenDTO tokenDTO = multiService.checkToken(accessToken);
+        if (tokenDTO.isOK()) try {
             multiService.deleteChatroom(chatroomId);
             return ResponseEntity.status(HttpStatus.OK).body(null);
             // throw new DataNotFoundException("not found chatroom");
         } catch (DataNotFoundException ex) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).body(ex.getMessage());
         }
+        else return tokenDTO.getResponseEntity();
     }
 }
